requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
axe-selenium-python==2.1.6
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
ConfigArgParse==1.7
cryptography==42.0.5
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
google-search-results==2.4.2
googlemaps==4.10.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
image-quality==1.2.7
imageio==2.34.0
importlib-metadata==7.1.0
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.0
keras==2.13.1
langdetect==1.0.9
language-detector==5.0.2
lazy-loader==0.4
libclang==18.1.1
libretranslatepy==2.1.1
libsvm==3.23.0.4
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pluggy==1.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic-core==2.16.3
pyLanguagetool==0.10.0
pymongo==4.6.3
pyphen==0.14.0
PySocks==1.7.1
pyspellchecker==0.8.1
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-whois==0.9.3
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
resmem==1.1.6
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
selenium
Selenium-Screenshot==2.1.0
sentence-transformers==2.6.1
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.37.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
textblob==0.18.0.post0
textstat==0.7.3
threadpoolctl==3.4.0
tifffile==2023.7.10
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
translate==3.6.1
trio==0.25.0
trio-websocket==0.11.1
triton==2.2.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
webdriver-manager==4.0.1
werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
zipp==3.18.1
